requirements.txt
streamlit
google-generativeai
duckduckgo-search
python-dotenv
gTTS
streamlit-mic-recorder
ffmpeg-python
ddgs